try to get connmon to listen to collectd directly
we want connmon to receive messages straight from a collectd
server that is getting messages from other collectd servers.
this means connmon has to report based on the Values() objects
passed in from collectd.   two problems:  1. these Values are broken
up for individual fields in each Type, so have to be re-aggregated,
or interpreted differently by the connmon client and 2. the collectd
server does not receive number of procs information; the values
are pre-aggreagted across all pids, so again this needs to be added
as a stat if connmon is to show it.   all in all a lot of deep thinking
has to happen to make this work.

Change-Id: I83dee419be15f7042175dbae7a79b40df1410037