Send zero to collectd process for interval
Observing the openstack confs have 120 seconds (!) for
the interval.   By sending interval as zero when we
do dispatch(), that means the data will be sent with the
configured interval.  This means for us, the correct
interval that the data is actually sent at will be present
in the data, so that in particular the connmon tool will know
more accurately how often data is being sent.

Change-Id: I42974904ee9e4c576e0aa5a0c04ab3886c19da9f