and now for something completely different
establish that we don't need to pull older values from
TimeBucket.   rewrite the TimeBucket concept completely so that
it strictly stores one set of values, each of which include its
own timestamp and interval, and simply blow away any keys that
are stale based on those numbers plus the incoming timestamp.

Assert that the time being queried / put with TimeBucket is
always growing.

Change-Id: I4a5dcf222a950dc10b0569ede7ed90c5361e786c