Refactor all systems to use Values; connmon becomes separate
The "connmon" tool can now display stats for SQLAlchemy stats that are sent
to a collectd server via the "network" plugin or through any other means.
Previously, the connmon tool could only display stats for messages that
were sent to a the collectd server by the SQLAlchemy-collectd plugin
itself.

The server side configuration for "connmon" is now separate from that of
the SQLAlchemy-collectd plugin, and the two plugins can run independently
of each other.  This allows for a configuration where many hosts send
SQLAlchemy-collectd messages to local collectd servers for aggregation, and
those servers then pass their messages onto another collectd server, where
the "connmon" tool can provide a view inside the current stats.

In order to achieve this, major refactoring such that the internals now
deal with data in terms of a structure which mirrors the collectd-python
"Values" object is in place, along with a rearchitecture of the connmon
tool such that it now consumes collectd "Values" objects from a particular
collectd server regardless of how those "Values" arrived in that server.

Fixes: #8
Change-Id: Idbbf607ad571f9eb723f2230ec1a3adad33aa536