reorganize networking and execution models to asyncio
this converts all networking other than the SQLAlchemy plugin
itself to use asyncio with UDP datagram protocols.   In addition,
the "connmon" plugin and client both gain the ability to either
be listeners or connectors, from a network connection point of
view.  Either many "connmon plugins" within collectd can all connect
into a single connmon client, or many connmon clients can connect
into a single "connmon plugin".

This is an interim step, where the "connmon plugin" will undergo
more changes:

1. the first is that it will known as a "relay", where we can
set up these relays in any configuration to bounce along to
each other in any configuration.  this will supersede the use of the
collectd "network" plugin to do the same thing.
2. the "connmon plugin", now the "relay", will be available as
its own standalone service.  the "collectd plugin" version of it will
be just a particular deployment option.
3. then we stick prometheus endpoints into the "relay server".

at that point, the "collectd" elements of the application will be
entirely contained in a "collectd" package and the "collectd" python
module will be only an optional dependency.

sqlalchemy-collectd will continue to use the collectd wire protocol
internally as it does already, as this is nicely set up for UDP packets
to fly in any direction.

Change-Id: I1a23f629f3a85fea414eda442d9d9fbe7f930d9d